+++ tilavarauspalvelu/api/graphql/types/reservation/serializers/deny_serializers.py
@@ -58,7 +58,7 @@ class ReservationDenySerializer(NestingModelSerializer):
         return data
 
     def update(self, instance: Reservation, validated_data: ReservationDenyData) -> Reservation:
-        if instance.access_type == AccessType.ACCESS_CODE:
+        if instance.access_type == AccessType.ACCESS_CODE and instance.recurring_reservation is None:
             with suppress(PindoraNotFoundError):
                 PindoraClient.delete_reservation(reservation=instance)
                 validated_data["access_code_generated_at"] = None
